+++ stackstate_checks_base/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 6.11.0 / 2020-03-09
+
+* [Added] Added support to map user defined `stackstate-environment` tags or config to the `environments` object
+* [Added] Added support to map user defined `stackstate-layer` tags or config to the `layer` object
+* [Added] Added support to map user defined `stackstate-domain` tags or config to the `domain` object
+* [Added] Added support to map user defined `stackstate-identifiers` tags or config to the `identifiers` array
 
 ## Apply patches up to 6.10.1 / 2019-07-09